+++ src/features/auth/ui/index.ts
@@ -1,4 +1,4 @@
 import LoginLayout from './LoginLayout';
-import SignUpLayout from './SignUpLayout';
+import SignUpLayout from './signup/SignUpLayout';
 import PasswordRequestLayout from './PasswordRequestLayout';
 export { LoginLayout, SignUpLayout, PasswordRequestLayout };
